+++ x-pack/platform/plugins/private/reporting/server/services/audit_events/audit_events.ts
@@ -32,7 +32,7 @@ const scheduledReportEventTypes: Record<
   scheduled_report_disable: 'change',
 };
 
-export interface ScheduledReportAuditEventParams {
+interface ScheduledReportAuditEventParams {
   action: ScheduledReportAuditAction;
   outcome?: EcsEvent['outcome'];
   savedObject?: NonNullable<AuditEvent['kibana']>['saved_object'];
