+++ x-pack/platform/plugins/private/reporting/server/services/scheduled_reports/index.ts
@@ -1,0 +1,10 @@
+/*
+ * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
+ * or more contributor license agreements. Licensed under the Elastic License
+ * 2.0; you may not use this file except in compliance with the Elastic License
+ * 2.0.
+ */
+
+export { scheduledQueryFactory } from './scheduled_query';
+
+export { DEFAULT_SCHEDULED_REPORT_LIST_SIZE, MAX_SCHEDULED_REPORT_LIST_SIZE } from './constants';
